+++ src/routing/ferry.fixture.ts
@@ -80,7 +80,7 @@ export function buildGraph(nodes: NodeSpec[], edges: EdgeSpec[]): RoutingGraph {
   const edgeCover = new Uint8Array(edgeCount);
   const edgeKindSide = new Uint8Array(edgeCount);
   const edgeFlags = new Uint8Array(edgeCount);
-  const edgeDurationSeconds = new Float32Array(edgeCount);
+  const edgeDurationSeconds = new Uint16Array(edgeCount);
   const edgeNameId = new Uint16Array(edgeCount).fill(NAME_NONE);
   const edgeGeomOffset = new Uint32Array(edgeCount).fill(NO_GEOMETRY);
   const edgeGeomCount = new Uint16Array(edgeCount);
